organize title bar folder
